Split unused file and tag entity deletion into a SELECT ... FOR UPDATE and a DELETE to prevent database locking

+++ genie-core/src/main/java/com/netflix/genie/core/services/TagService.java
@@ -47,5 +47,5 @@ public interface TagService {
      *                         will be deleted. Inclusive
      * @return The number of tags deleted
      */
-    int deleteUnusedTags(@NotNull final Instant createdThreshold);
+    long deleteUnusedTags(@NotNull final Instant createdThreshold);
 }
